+++ src/fr/maxlego08/menu/loader/permissible/CurrencyPermissibleLoader.java
@@ -5,16 +5,18 @@ import fr.maxlego08.menu.api.requirement.Action;
 import fr.maxlego08.menu.api.requirement.Permissible;
 import fr.maxlego08.menu.api.utils.TypedMapAccessor;
 import fr.maxlego08.menu.loader.ZPermissibleLoader;
-import fr.maxlego08.menu.requirement.permissible.ZVaultPermissible;
+import fr.maxlego08.menu.requirement.permissible.ZCurrencyPermissible;
+import fr.traqueur.currencies.Currencies;
 
 import java.io.File;
+import java.math.BigDecimal;
 import java.util.List;
 
-public class VaultPermissibleLoader extends ZPermissibleLoader {
+public class CurrencyPermissibleLoader extends ZPermissibleLoader {
 
     private final ButtonManager buttonManager;
 
-    public VaultPermissibleLoader(ButtonManager buttonManager) {
+    public CurrencyPermissibleLoader(ButtonManager buttonManager) {
         this.buttonManager = buttonManager;
     }
 
@@ -27,6 +29,9 @@ public class VaultPermissibleLoader extends ZPermissibleLoader {
     public Permissible load(String path, TypedMapAccessor accessor, File file) {
         List<Action> denyActions = loadAction(buttonManager, accessor, "deny", path, file);
         List<Action> successActions = loadAction(buttonManager, accessor, "success", path, file);
-        return new ZVaultPermissible(denyActions, successActions, accessor.getDouble("amount"));
+        String amount = accessor.getString("amount");
+        Currencies currencies = Currencies.valueOf(accessor.getString("currency", Currencies.VAULT.name()));
+        String economyName = accessor.getString("economy", null);
+        return new ZCurrencyPermissible(denyActions, successActions, currencies, amount, economyName);
     }
 }